==> rmd/04_convert_pdfs_to_pngs.Rmd <==
---
title: "convert_to_png"
output: html_document
---

# Convert PDFs to PNGs


```{r prep}
set.seed(2)

required_packages <- c("tidyverse", "magrittr", "DBI", "bigrquery", "arrow","glue", "vroom","janitor", "gt", "ggwordcloud", "readxl", "ggthemes", "hrbrthemes", "extrafont", "plotly", "scales", "stringr", "gganimate", "here", "tidytext", "sentimentr", "scales", "DT", "here", "sm", "mblm", "glue", "fs", "knitr", "rmdformats", "janitor", "urltools", "colorspace", "pdftools", "showtext", "pander", "ggridges", "spatstat", "broom")
for(i in required_packages) { 
  if(!require(i, character.only = T)) {
    #  if package is not existing, install then load the package
    install.packages(i, dependencies = T)
  require(i, character.only = T)
  }
}
```



```{r convert-pdf-to-png, echo = F}
pdfs <- list.files(here("plots"), pattern = "*.pdf")
                                      setwd(here("plots", "png"))
for(pdf in pdfs) {
  pdf_convert(pdf = here("plots", pdf),
              format = "png", dpi = 720)
}
# pdf <- "3_1_jobs_cat2.pdf"
pdf_convert(pdf = here("plots", "5_1_require_edu_size_n.pdf"),  format = "png", dpi = 720)
```



<!-- ```{r convert_to_pngs, results='asis', echo=FALSE} -->
<!-- dir_plots <- here::here("plots") -->

<!-- pdfs <- list.files(dir_plots, pattern = "map.*pdf", recursive = TRUE) -->
<!-- for(pdf in pdfs) { -->
<!--   pdf_convert(pdf = glue::glue("{dir_plots}/{pdf}"),  -->
<!--               filenames = glue::glue("{dir_plots}/png/{str_remove(pdf, '.pdf')}.png"), -->
<!--               format = "png", dpi = dpi) -->
<!-- } -->
<!-- ```   -->
